Fix misreported sensor status
Sensor status can be misreported as `Online` due to a values that were not originally anticipated.

The logic that assigns the `Online` tag simply checks to see if the `status` field exists. However, there are cases where the field can exist but not be the value `Online`. I have identified an edge case where the status can be `Uninstall Pending`.

To fix this issue, I have added an additional check to ensure that the status exists and is `Online`.